changed structure to allow multipage apps
still having problem with callbacks, working.
