+++ pages/cicd.py
@@ -1,4 +1,5 @@
 from dash import html
+from app import app
 
 cicd_layout = html.Div(children=[
     html.H1(children="CI/CD Page!")
